fix/score-to-integer-mapping (#30)
* fix: change rating scores from Double to Integer in API and Mapper

- This aligns the code with the openAPI specification.
- Update RatingReadDto fields (quality, cost, reliability, availability) to Integer.
- Implement parseIntegerOrNull utility to handle OpenBIS float-string formats (e.g. "1.0" -> 1).
- Refactor OpenBisUtils to reuse error logging logic.
- Fixes failing Unit test case in RatingMapperTest.

* docs: clarify rounding behavior in RatingMapper and OpenBisUtils

- Add comments explaining the rounding of floating-point rating properties to integers within RatingMapper.
- Document the rounding logic in the parseIntegerOrNull method of OpenBisUtils.

* docs: fix typo in comments and remove redundant notes in RatingMapper

- Correct "which where" to "which were" in comments.
- Remove repetitive comments referencing QUALITY_RATING_PROPERTY.

+++ src/main/java/io/github/supplierratingsoftware/supplierratingbackend/dto/api/RatingReadDto.java
@@ -41,16 +41,16 @@ package io.github.supplierratingsoftware.supplierratingbackend.dto.api;
  * @param supplierName The supplier name (Parent of the order).
  */
 public record RatingReadDto(
-        Double quality,
+        Integer quality,
         String qualityReason,
 
-        Double cost,
+        Integer cost,
         String costReason,
 
-        Double reliability,
+        Integer reliability,
         String reliabilityReason,
 
-        Double availability,
+        Integer availability,
         String availabilityReason,
 
         Double totalScore,
